requirements.txt
# Core ML Libraries
torch>=2.0.0
transformers>=4.36.0
peft>=0.7.0
bitsandbytes>=0.41.0
accelerate>=0.25.0
datasets>=2.16.0
scipy>=1.11.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Evaluation
scikit-learn>=1.3.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
ruff>=0.1.0
mypy>=1.7.0

# Jupyter (optional)
jupyter>=1.0.0
ipywidgets>=8.1.0
